authorization on edit and delete posts

+++ baackend/routes/posts.js
@@ -34,7 +34,7 @@ const storage = multer.diskStorage({
 
 //const upload = multer({ dest: 'baackend/images/'});
 
-// upload image
+// create post
 router.post('', multer({ storage: storage }).single('image'), 
     checkAuth,
     (req, res, next) => {
@@ -42,8 +42,10 @@ router.post('', multer({ storage: storage }).single('image'),
     const post = new Post({
         title: req.body.title,
         content: req.body.content,
-        imagePath: url + '/images/' + req.file.filename
+        imagePath: url + '/images/' + req.file.filename,
+        creator: req.userData.userId 
     });
+
     post.save().then((createdPost) => {
         res.status(200).json({
             message: 'post is send',
@@ -52,7 +54,7 @@ router.post('', multer({ storage: storage }).single('image'),
                 title: createdPost.title,
                 content: createdPost.content,
                 imagePath: createdPost.imagePath,
-
+               creator: createdPost.creator 
             }
         })
     })
@@ -77,14 +79,18 @@ router.put('/:id',
         _id: req.body.id,
         title: req.body.title,
         content: req.body.content,
-        imagePath : imagePath
+        imagePath : imagePath,
+        creator: req.userData.userId
     })
 
-     console.log(post);
+    // console.log(post);
 
-    Post.updateOne({ _id: req.params.id }, post).then(result => {
-        console.log(result);
-        res.status(200).json({ message: "Update successful!" });
+    Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post).then(result => {
+        if (result.nModified > 0) {
+            res.status(200).json({ message: "Update successful!" });
+        }else{
+            res.status(401).json({ message: "Update Failed unauthorized!" });
+        }
     });
 });
 
@@ -135,12 +141,12 @@ router.get('', (req, res, next) => {
 router.delete('/:id',
     checkAuth,
      (req, res, next) => {
-    Post.deleteOne({ _id: req.params.id }).then(result => {
-        console.log(result);
-
-        res.status(200).json({
-            message: 'post deleted',
-        })
+    Post.deleteOne({ _id: req.params.id ,  creator: req.userData.userId}).then(result => {
+        if (result.n > 0) {
+            res.status(200).json({ message: "deletion successful!" });
+        }else{
+            res.status(401).json({ message: "deletion Failed unauthorized!" });
+        }
     })
 
 });
